refactor: rename tokens.js to messages.js
