+++ lib/services/__tests__/account.test.js
@@ -3,7 +3,7 @@ import Config from 'react-native-config'
 import * as uuid from 'uuid'
 import * as crypto from '../crypto'
 import * as storage from '../storage'
-import * as tokens from '../tokens'
+import * as messages from '../messages'
 import { generateTestAccount } from './_helpers'
 
 jest.mock('uuid', () => ({
@@ -23,10 +23,10 @@ jest.mock('../storage', () => ({
     .mockName('storage.storeKey')
     .mockResolvedValue(),
 }))
-jest.mock('../tokens', () => ({
+jest.mock('../messages', () => ({
   createAccountRegistration: jest
     .fn()
-    .mockName('tokens.createAccountRegistration'),
+    .mockName('messages.createAccountRegistration'),
 }))
 
 Config.OPERATOR_URL = 'aTotallyLegitOperatorUrl'
@@ -67,7 +67,7 @@ describe('account', () => {
         privateKeyPem,
       })
       crypto.toPublicKey.mockReturnValue(publicKey)
-      tokens.createAccountRegistration.mockResolvedValue(jwt)
+      messages.createAccountRegistration.mockResolvedValue(jwt)
       fetch.mockResponse(JSON.stringify({ data: { id: 'abc123' } }))
     })
     describe('new user', () => {
@@ -95,11 +95,14 @@ describe('account', () => {
       })
       it('generates a token', async () => {
         const result = await accountService.save(account)
-        expect(tokens.createAccountRegistration).toHaveBeenCalledWith(result, {
-          publicKey,
-          privateKey,
-          privateKeyPem,
-        })
+        expect(messages.createAccountRegistration).toHaveBeenCalledWith(
+          result,
+          {
+            publicKey,
+            privateKey,
+            privateKeyPem,
+          },
+        )
       })
       it('saves account to storage', async () => {
         const result = await accountService.save(account)
