+++ lib/services/__tests__/auth.test.js
@@ -7,13 +7,13 @@ import { getAccount } from '../account'
 import { generateTestAccount } from './_helpers'
 import { generateKey } from '../crypto'
 import Config from 'react-native-config' // this is mocked
-import * as tokens from '../tokens'
+import * as messages from '../messages'
 
 jest.mock('../account', () => ({
   getAccount: jest.fn(),
 }))
 
-jest.mock('../tokens', () => ({
+jest.mock('../messages', () => ({
   createConnectionInit: jest
     .fn()
     .mockResolvedValue('jwt-for-createConnectionInit'),
@@ -256,16 +256,16 @@ describe('auth', () => {
       afterEach(() => {
         generateKey.mockReset()
       })
-      it('calls tokens.createConnection with correct arguments for empty approved', async () => {
+      it('calls messages.createConnection with correct arguments for empty approved', async () => {
         await approveConnection(connectionRequest, new Map())
 
-        expect(tokens.createConnection).toHaveBeenCalledWith(
+        expect(messages.createConnection).toHaveBeenCalledWith(
           connectionRequest,
           undefined,
           expect.any(String),
         )
       })
-      it('calls tokens.createConnection with correct arguments, some approved', async () => {
+      it('calls messages.createConnection with correct arguments, some approved', async () => {
         const approved = new Map([
           ['18710e28-7d6c-49cf-941e-0f954bb179ae', true],
           ['1712ec0c-9ae6-472f-9e14-46088e51f505', true],
@@ -275,7 +275,7 @@ describe('auth', () => {
         ])
         await approveConnection(connectionRequest, approved)
 
-        expect(tokens.createConnection).toHaveBeenCalledWith(
+        expect(messages.createConnection).toHaveBeenCalledWith(
           connectionRequest,
           {
             approved: expect.any(Array),
@@ -283,10 +283,10 @@ describe('auth', () => {
           expect.any(String),
         )
 
-        const permissions = tokens.createConnection.mock.calls[0][1]
+        const permissions = messages.createConnection.mock.calls[0][1]
         expect(permissions.approved).toHaveLength(4)
       })
-      it('calls tokens.createConnection with correct arguments, none approved', async () => {
+      it('calls messages.createConnection with correct arguments, none approved', async () => {
         const approved = new Map([
           ['18710e28-7d6c-49cf-941e-0f954bb179ae', false],
           ['1712ec0c-9ae6-472f-9e14-46088e51f505', false],
@@ -296,7 +296,7 @@ describe('auth', () => {
         ])
         await approveConnection(connectionRequest, approved)
 
-        expect(tokens.createConnection).toHaveBeenCalledWith(
+        expect(messages.createConnection).toHaveBeenCalledWith(
           connectionRequest,
           undefined,
           expect.any(String),
