+++ lib/services/auth.js
@@ -7,7 +7,7 @@ import {
   createConnectionResponse,
   createLogin,
   createLoginResponse,
-} from './tokens'
+} from './messages'
 import { v4 } from 'uuid'
 import { generateKey, toPublicKey } from './crypto'
 
@@ -30,9 +30,9 @@ export const initConnection = async authRequest => {
       headers: { 'content-type': 'application/jwt' },
       body: connectionInit,
     })
-    const data = await response.text()
-    const { payload } = await verify(data)
-    return payload
+    const JWTConnectionRequest = await response.text()
+    const { payload: connectionRequest } = await verify(JWTConnectionRequest)
+    return connectionRequest
   } catch (error) {
     console.error(error)
     throw Error('CONNECTION_INIT failed')
